requirements.txt
google-api-python-client==1.6.5
google-auth==1.4.1
google-auth-httplib2==0.0.3
networkx==2.1
nltk==3.2.5
numpy==1.14.2